data.py
The code above replaces the data.py in the datasets folder. It would be good to have this some reverse-normalizer. Class "WasteWaterData" for analysis of the analysis of Ontario Ministry of Environment, Conservation and Parks data on wastewater.
Ontario Contains multiple public health units (PHUs), within each PHU's there are multiple locations (latitude and longitude) that takes the measurement of virulence factors measured as mN1, mN2, mQ, mBiomarkers and geographical measurements. The class WasteWaterData picks up relevant lat-long as well as other relevant columsn needed. There are some un-necessary functions which are solely for personal use and can be amended as per your needs.
Thanks,
Jag